+++ src/Sparrow/Platform/Posix/PosixUnmanagedMemory.cs
@@ -10,18 +10,18 @@ namespace Sparrow
         
         [DllImport(LIBC_6, EntryPoint = "memcpy", CallingConvention = CallingConvention.Cdecl, CharSet = CharSet.Unicode, SetLastError = false)]
         [SecurityCritical]
-        public static extern IntPtr Copy(byte* dest, byte* src, int count);
+        public static extern IntPtr Copy(byte* dest, byte* src, long count);
 
         [DllImport(LIBC_6, EntryPoint = "memcmp", CallingConvention = CallingConvention.Cdecl, SetLastError = false)]
         [SecurityCritical]
-        public static extern int Compare(byte* b1, byte* b2, int count);
+        public static extern int Compare(byte* b1, byte* b2, long count);
 
         [DllImport(LIBC_6, EntryPoint = "memmove", CallingConvention = CallingConvention.Cdecl, SetLastError = false)]
         [SecurityCritical]
-        public static extern int Move(byte* dest, byte* src, int count);
+        public static extern int Move(byte* dest, byte* src, long count);
 
         [DllImport(LIBC_6, EntryPoint = "memset", CallingConvention = CallingConvention.Cdecl, SetLastError = false)]
         [SecurityCritical]
-        public static extern IntPtr Set(byte* dest, int c, int count);
+        public static extern IntPtr Set(byte* dest, int c, long count);
     }
 }
